feat(FunctionType): functionNumber로부터 해당하는 enum값 리턴 구현
해당하는 값 없으면 IllegalArgumentException 에러 발생
